chore(ui): move FilterList component to shared
